pc/sio: Note output error even in non-verbose mode; clean up debugging
If write() returns less than the full value we expected to write this is a
serious problem and the transfer is definitely going to fail, so print out
the "sio::Write()" error message even if we're not in verbose mode.

BUG: Actually, we should retry the remainder of the write at least once, in
case we overflowed buffers or whatever, but that's a much bigger change
than I care to get in to right now.

Also, re-enable the verbose debugging output (activated by the -v flag),
make it more clear in the message that this output also includes error
information, and remove other commented-out debugging prints that are not
useful in light of this.